requirements.txt
Django==4.0
psycopg2_binary==2.9.9
djangorestframework==3.14.0
django-filter==23.3
djangorestframework-simplejwt==5.3.0
drf-nested-routers==0.93.4